+++ tests/cpu_checking/static/test_warnings/main.cpp
@@ -6,8 +6,8 @@
 int main()
 {
   int n = 8;
-  int nbytes = n*sizeof(double); 
-  double *data = (double *)malloc(nbytes);
+  int nbytes = n*((int)sizeof(double)); 
+  double *data = (double *)malloc((size_t)nbytes);
   for (int i=0; i < n; ++i)
     data[i] = (double)(i+1);
   printf("Calling kernel\n");
